% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heat_index.r
\name{heat_index}
\alias{heat_index}
\title{heat_index}
\usage{
heat_index(t, rh)
}
\arguments{
\item{numeric}{t Air temperature in Celsius Degrees.}

\item{numeric}{rh Air Relative humidity .}
}
\value{

}
\description{
Calculate the heat index following NOAA scheme.
}
\author{
Istituto di Biometeorologia Firenze Italy  Alfonso Crisci \email{a.crisci@ibimet.cnr.it}
}
\references{
NOAA index from George Winterling based on SteadMan's work.
}
\keyword{NOAA}
\keyword{heat}
\keyword{index}

